+++ lua/plugins/config/notify.lua
@@ -1,0 +1,8 @@
+return function(use)
+	use({
+		'rcarriga/nvim-notify',
+		config = function()
+			vim.notify = require('notify')
+		end,
+	})
+end