+++ lua/plugins/config/tmux_config.lua
@@ -14,10 +14,6 @@ return function(use)
 					-- enables default keybindings (C-hjkl) for normal mode
 					enable_default_keybindings = true,
 				},
-				resize = {
-					-- enables default keybindings (A-hjkl) for normal mode
-					enable_default_keybindings = true,
-				},
 			})
 		end,
 	})