+++ lua/plugins/config/golang_enhancement.lua
@@ -13,22 +13,12 @@ return function(use)
 					end)
 
 					if not gopls:is_installed() then
+						vim.notify('installing gopls...')
 						gopls:install()
 					end
 				end
 				require('go').setup({
 					gofmt = 'gofumpt',
-					max_line_len = 120, -- max line length in goline format
-					tag_transform = false, -- tag_transfer  check gomodifytags for details
-					verbose = false, -- output loginf in messages
-					lsp_cfg = false, -- true: apply go.nvim non-default gopls setup
-					lsp_gofumpt = true, -- true: set default gofmt in gopls format to gofumpt
-					lsp_on_attach = require('plugins.config.lsp.on_attach'),
-					lsp_diag_hdlr = true, -- hook lsp diag handler
-					dap_debug = false, -- set to true to enable dap
-					dap_debug_keymap = true, -- set keymaps for debugger
-					dap_debug_gui = true, -- set to true to enable dap gui, highly recommended
-					dap_debug_vt = true, -- set to true to enable dap virtual text
 				})
 			end,
 			ft = { 'go' },