+++ lua/plugins/config/tmux_config.lua
@@ -8,7 +8,7 @@ return function(use)
 				copy_sync = {
 					-- enables copy sync and overwrites all register actions to
 					-- sync registers *, +, unnamed, and 0 till 9 from tmux in advance
-					enable = true,
+					enable = false,
 				},
 				navigation = {
 					-- enables default keybindings (C-hjkl) for normal mode