+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased]
 
 ### Bug Fixes
+- **lsp-status:** now lsp status line does not check buf get clients
 - **prettierd:** fix prettierd config for null_ls
 - **tmux:** removed resize default keybinding because conflict with line swap
 
@@ -13,12 +14,13 @@
 
 ### Features
 - **cmp-lspkind:** update configuration
+- **hlargs:** added
 - **lsp:** update lsp in insert mode disabled to increase peformance on slow lsps
 - **lualine:** better color for modified
 - **mini:** disabled on some filetypes
 - **mini-indentscope:** added mini indentscope
-- **notify:** added nvim-notify
 - **notify:** added notify on installing sumneko lua server
+- **notify:** added nvim-notify
 - **null_ls:** revert to prettierd and eslint_d now only runs if eslintrc.json exist in root
 - **pairs:** moved to smart pairs
 - **statusline:** revert to lualine from windline