+++ CHANGELOG.md
@@ -17,17 +17,18 @@
 - **cmp-lspkind:** update configuration
 - **golang:** update enhancement configuration
 - **hlargs:** added
-- **lsp:** fix cursor jumpt to diagnostic window on goto_next
 - **lsp:** update lsp in insert mode disabled to increase peformance on slow lsps
+- **lsp:** fix cursor jumpt to diagnostic window on goto_next
 - **lualine:** better color for modified
 - **mini:** disabled on some filetypes
 - **mini-indentscope:** added mini indentscope
-- **notify:** added nvim-notify
 - **notify:** added notify on installing sumneko lua server
+- **notify:** added nvim-notify
 - **null_ls:** revert to prettierd and eslint_d now only runs if eslintrc.json exist in root
 - **pairs:** moved to smart pairs
-- **sandwhich:** fix missing mappings
 - **sandwhich:** moved from surround.nvim to vim-sandwhich because dev account is deleted
+- **sandwhich:** fix missing mappings
+- **session_manager:** session manager now autoload currentdir
 - **statusline:** revert to lualine from windline
 - **tmux:** disabled copy_sync to sync clipboard with system clipboard instead of tmux
 - **tmux:** added tmux config